feat: add export command and modularize CLI
