+++ tests/unit/test_save_naming.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from tushare_a_fundamentals import cli as appmod
+from tushare_a_fundamentals.common import save_tables
 
 pytestmark = pytest.mark.unit
 
@@ -11,5 +11,5 @@ def test_save_naming(tmp_path):
         "raw": pd.DataFrame({"ts_code": ["000001.SZ"], "end_date": ["20231231"]}),
     }
     outdir = tmp_path
-    appmod.save_tables(tables, str(outdir), "income_vip_quarter", "csv")
+    save_tables(tables, str(outdir), "income_vip_quarter", "csv")
     assert (outdir / "csv" / "income_vip_quarter_raw.csv").exists()
